Only perform input retrieval for testnet
